oops... some more 'hides' warnings
git-svn-id: http://svn.miranda-ng.org/main/trunk@15100 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ plugins/Toaster/src/string_reference_wrapper.h
@@ -25,7 +25,7 @@ public:
 	template <size_t _>
 	StringReferenceWrapper(_In_reads_(_) wchar_t(&stringRef)[_]) throw()
 	{
-		UINT32 length = wcslen(stringRef);
+		size_t length = wcslen(stringRef);
 		WindowsCreateStringReference(stringRef, length, &_header, &_hstring);
 	}
